feature: conexion a api externa baja performance

+++ src/main/java/ar/edu/unq/desapp/grupoE/backenddesappapi/model/CryptoQuote.java
@@ -4,10 +4,10 @@ import java.time.LocalDateTime;
 
 public class CryptoQuote {
     private String name;
-    private Float price;
+    private Double price;
     private LocalDateTime updateTime;
 
-    public CryptoQuote(String name, Float price, LocalDateTime updateTime) {
+    public CryptoQuote(String name, Double price, LocalDateTime updateTime) {
         this.name = name;
         this.price = price;
         this.updateTime = updateTime;
@@ -17,7 +17,7 @@ public class CryptoQuote {
         return name;
     }
 
-    public Float getPrice() {
+    public Double getPrice() {
         return price;
     }
